+++ src/lib/types.ts
@@ -83,6 +83,18 @@ export interface Extend {
     meta?: DefinitionMeta,
     ota?: DefinitionOta,
     onEvent?: OnEvent,
+    isModernExtend?: false,
+}
+
+export interface ModernExtend {
+    fromZigbee?: Fz.Converter[],
+    toZigbee?: Tz.Converter[],
+    exposes?: Expose[],
+    configure?: Configure,
+    meta?: DefinitionMeta,
+    ota?: DefinitionOta,
+    onEvent?: OnEvent,
+    isModernExtend: true,
 }
 
 export interface OnEventData {
@@ -110,7 +122,7 @@ export type Definition = {
     onEvent?: OnEvent,
     ota?: DefinitionOta,
 } & ({ zigbeeModel: string[] } | { fingerprint: Fingerprint[] })
-    & ({ extend: Extend, fromZigbee?: Fz.Converter[], toZigbee?: Tz.Converter[],
+    & ({ extend: Extend | ModernExtend[], fromZigbee?: Fz.Converter[], toZigbee?: Tz.Converter[],
         exposes?: (Expose[] | ((device: Zh.Device, options: KeyValue) => Expose[])) } |
     { fromZigbee: Fz.Converter[], toZigbee: Tz.Converter[], exposes: (Expose[] | ((device: Zh.Device, options: KeyValue) => Expose[])) });